+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-03T13:54:06.303Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-04T07:59:32.361Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")