+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGHWGRBM8SP7HXT35DT7CNS7"
+        "01KGKTM7DEKK188CHBVZ6BSA1N"
     )
     r"""The unique identifier of the data source"""